requirements.txt
deep_translator==1.11.4
Flask==3.1.2
flask_cors==6.0.2
gdown==5.2.1
nltk==3.9.2
numpy==2.4.1
Pillow==12.1.0
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
transformers==4.57.5
sentencepiece==0.2.1